requirements.txt
pyspark
pysimdjson
fsspec
pandas
loguru
pyarrow